Moved Lamar to use JasperFx, folded Lamar.Diagnostics into Lamar itself
