+++ src/Lamar/CommandLine/LamarServicesInput.cs
@@ -1,13 +1,13 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using JasperFx.CommandLine;
 using JasperFx.Core;
 using Lamar.IoC.Diagnostics;
-using Oakton;
 
-[assembly:OaktonCommandAssembly]
+[assembly:JasperFxAssembly]
 
-namespace Lamar.Diagnostics
+namespace Lamar.CommandLine
 {
     
     public class LamarServicesInput : NetCoreInput
