+++ src/Lamar/CommandLine/ServiceCollectionExtensions.cs
@@ -1,8 +1,8 @@
 using JasperFx.Core.TypeScanning;
+using JasperFx.Environment;
 using Microsoft.Extensions.DependencyInjection;
-using Oakton.Environment;
 
-namespace Lamar.Diagnostics
+namespace Lamar.CommandLine
 {
     public static class ServiceCollectionExtensions
     {
