+++ src/Lamar/CommandLine/ValidateInput.cs
@@ -1,6 +1,6 @@
-using Oakton;
+using JasperFx.CommandLine;
 
-namespace Lamar.Diagnostics
+namespace Lamar.CommandLine
 {
     public class ValidateInput : NetCoreInput
     {
