--- src/Lamar/IgnoreAssemblyAttribute.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Lamar;
-
-[AttributeUsage(AttributeTargets.Assembly)]
-public class IgnoreAssemblyAttribute : Attribute
-{
-}
