+++ src/Lamar/IoC/Frames/GetInstanceFrame.cs
@@ -46,7 +46,7 @@ public class GetInstanceFrame : SyncFrame, IResolverFrame
         if (Next is null)
         {
             throw new InvalidCastException(
-                $"{typeof(GetInstanceFrame).GetFullName()}.{nameof(Next)} must not be null.");
+                $"{typeof(GetInstanceFrame).FullNameInCode()}.{nameof(Next)} must not be null.");
         }
     }
